refactor: немного поменял структуру папок и файлов
