--- packages/server/api/auth/index.ts
@@ -1,8 +1,0 @@
-import {yandexProxyAll, yandexProxyUserInfoOnly} from './authorization';
-import {yandexCheckAuthorization} from './checking';
-
-export {
-	yandexProxyAll,
-	yandexProxyUserInfoOnly,
-	yandexCheckAuthorization,
-};
