+++ packages/server/api/models/index.ts
@@ -1,5 +1,7 @@
 import {Users} from './users';
+import type {TUser} from './users';
 
 export {
 	Users,
+	TUser,
 };
