+++ packages/server/api/models/users.ts
@@ -3,7 +3,7 @@ import type {ModelAttributes} from 'sequelize/types';
 import {sequelize} from 'server/api/sequelize';
 
 // Модель таблицы Users
-type TUser = {
+export type TUser = {
 	id: number;
 	login: string;
 	nickname: string;
