+++ packages/server/api/user/index.ts
@@ -1,5 +1,5 @@
 import {Users} from 'server/api/models';
-import type {TUser} from 'server/api/auth/typing';
+import type {TUser} from 'server/api/models';
 import {dbConnect} from 'server/api/sequelize';
 
 export const userAPI = {
