--- packages/server/api/utils/index.ts
@@ -1,3 +1,0 @@
-import {yandexProxyResponseHandler} from 'server/api/utils/yandexProxyResponseHandler';
-
-export {yandexProxyResponseHandler};
