+++ packages/server/authMiddleware/utils/index.ts
@@ -1,0 +1,7 @@
+import {yandexProxyResponseHandler} from 'server/authMiddleware/utils/yandexProxyResponseHandler';
+import {filterCookies} from 'server/authMiddleware/utils/filterCookies';
+
+export {
+	yandexProxyResponseHandler,
+	filterCookies,
+};
